fixed the print function and moved the print function to utils
