+++ chemistry/chemistry_system/views.py
@@ -11,6 +11,7 @@ from django.views.decorators.http import require_POST
 from django.utils.timezone import now
 from django.core.paginator import Paginator
 from PIL import Image, ImageDraw, ImageFont
+from django.http import HttpResponse
 import qrcode
 import io
 import random
@@ -269,79 +270,10 @@ def print_page(request):
     """Render the page with the download button."""
     return render(request, 'print.html')
 
+    
 @login_required
-def generate_qr_pdf(request):
-    num_qr = 24  # Number of QR codes
-    cols = 4  # Number of columns
-    rows = 6   # Number of rows (num_qr / cols)
-    qr_size = 300  # Size of each QR code in pixels
-
-    # Calculate canvas size (including margins)
-    dpi = 300  
-    canvas_width = 8.5 * dpi  
-    canvas_height = 11 * dpi  
-    canvas_size = (int(canvas_width), int(canvas_height))
-
-    # Create a blank canvas
-    canvas = Image.new("RGB", canvas_size, "white")
-    draw = ImageDraw.Draw(canvas)
-
-    first_x = 384+40
-    first_y = 250
-    between_x = 565
-    between_y = 565
-    num_so_far = 0
-
-    current_x = first_x
-    current_y = first_y
-
-    current_directory = os.path.dirname(os.path.realpath(__file__))
-
-    font_path = os.path.join(current_directory, "tnr.ttf")  # Replace with your font file name
-    font_size = 24  # Adjust the font size here
-    font = ImageFont.truetype(font_path, font_size)
-    # Generate and place QR codes
-    for i in range(num_qr):
-        if num_so_far % cols == 0 and num_so_far != 0:
-            current_x = first_x
-            current_y += between_y
-        num_so_far += 1
-        data = random.randint(0, 12000000000)  # Unique data for each QR code
-        qr = qrcode.QRCode(box_size=5, border=0)  # Adjust size
-        qr.add_data(data)
-        qr.make(fit=True)
-        qr_img = qr.make_image(fill="black", back_color="white").resize((qr_size, qr_size))
-
-        # Compute position with margins
-        
-
-        # Paste QR code on canvas
-        canvas.paste(qr_img, (current_x-150, current_y-150))
-
-        # Draw text on canvas
-        position = (current_x-50, current_y + 175)
-        draw.text(position, data, font=font, fill="black")
-
-        # Update position for next QR code
-        current_x += between_x
-
-        
-
-    # trim the right side of the image
-    canvas = canvas.crop((0, 0, canvas_width, canvas_height))
-    # set image size to 9*11 inches
-    canvas = canvas.resize((int(9*300), int(11*300)))
-
-    pdf_buffer = io.BytesIO()
-    canvas.save(pdf_buffer, format="PDF")
-    pdf_buffer.seek(0)
-
-    # Return as a response
-    response = HttpResponse(pdf_buffer, content_type='application/pdf')
-    response['Content-Disposition'] = 'attachment; filename="qr_codes.pdf"'
-    return response
 def generate_qr_pdf_view(request):
-    return generate_qr_pdf()
+    return generate_qr_pdf(request)
 
 @login_required
 def log(request):
